Add image-only 'media' posts and style for homepage
To start, these just live in "posts" with everything else. The homepage
template checks for the 'media' category and renders them differently in
the homepage grid.

This feels a little hacky - the content is really different from the
other posts - but I don't have a very good sense of what these image
only posts are yet.