Prepend site.github.url to links to fix github pages deployment
The gh pages deploy lives at luncheonette.github.io/the-sunview/, which
broke a lot of links that worked locally.

Jekyll recommends this approach in their [Github Pages
docs](https://jekyllrb.com/docs/github-pages/). I don't love it, but the
other option is to set the baseurl config which means you have to
remember to remove that when running `jekyll serve` locally. Seems
harmless to include so far.