Limit # of events shown on homepage to 20
Avoids requiring the JS to run through all 100+ events and determine
whether they are upcoming or not. There are generally only a few
upcoming events listed at a time so 20 should be *more* than enough.

Refactor "_includes/events-list.html" so that it just includes the
markup for the list item. This means any page that needs to show events
can set up its own list of events (with limits).