"Pin" images and pages at certain positions in homepage grid
This is nasty, but liquid doesn't have any way I can find to just insert
something into an array at a certain index...

The homepage grabs the collection of all site documents that include the
homepage_position frontmatter attribute, and also assigns a
`pinned_counter` variable.

During each iteration of the for loop for displaying event posts, there
is a condition that uses the counter variable to see if the pinned doc
at its index has a homepage_position attribute equal to the current
index of the forloop. If it does, an additional element with the
"pinned" content is prepended.

It's also necessary to check the collection type in order to render the
correct markup for the pinned content.